fix: cast wgctrl int64 bytes to uint64 in WireGuard metrics
peer.ReceiveBytes and peer.TransmitBytes are int64 in wgctrl,
need explicit cast to uint64 for PeerDetail struct.

+++ internal/vpn/wireguard_metrics.go
@@ -57,8 +57,8 @@ func (c *WireGuardMetricsCollector) GetPeerMetrics() (*WireGuardPeerMetrics, err
 	for _, peer := range device.Peers {
 		detail := PeerDetail{
 			PublicKey:     peer.PublicKey.String(),
-			BytesReceived: peer.ReceiveBytes,
-			BytesSent:     peer.TransmitBytes,
+			BytesReceived: uint64(peer.ReceiveBytes),
+			BytesSent:     uint64(peer.TransmitBytes),
 			LastHandshake: peer.LastHandshakeTime,
 			AllowedIPs:    make([]string, len(peer.AllowedIPs)),
 		}
